+++ packages/toolbars-ui/web/src/buttons/DeleteButton/DeleteButton.tsx
@@ -11,10 +11,14 @@ interface Props {
 
 const DeleteButton: FC<Props> = ({ toolbarItem }) => {
   const { t } = useContext(RicosContext) || {};
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  const node = (toolbarItem.attributes as Record<string, any>).selectedNode;
+
+  const nodeId = node?.attrs.id;
   return (
     <ToggleButton
       Icon={TrashIcon}
-      onClick={toolbarItem.commands.delete}
+      onClick={() => toolbarItem.commands.delete(nodeId)}
       dataHook="blockButton_delete"
       tooltip={t('DeleteButton_Tooltip')}
     />