+++ CHANGELOG.md
@@ -21,6 +21,8 @@
 
 - `common-plugins`
   - [#4214](https://github.com/wix/ricos/pull/4214) common-plugins package
+- `ricos-plugins`
+  - [#4238](https://github.com/wix/ricos/pull/4238) plugin toolbar delete command by id instead of selection
 
 </details>